fix(integrators): step and controller build() write their products into the cache explicitly; plain callables compare by equality so the run passes system.get_solver_helper directly; the FIRK norm takes its coefficients at construction only
The product_field mechanism and the _build fill loop are gone: BaseAlgorithmStep.build() calls compile_step() and BaseStepController.build() calls compile_controller(), each recording its sizes, order, flags and bounds beside the device function. _nested_config_fields no longer filters on an update method (every nested config class has one). _values_differ compares only device-function fields by identity, so a rebinding of the system's helper getter is not a change and the run needs no bound-once attribute; the explicit Euler step stores the getter. The FIRK correction norm drops its stage-count validation and the step's update override that pushed coefficients on a tableau change. The controller-order tests build in float32.

+++ tests/integrators/algorithms/test_controller_order.py
@@ -61,7 +61,7 @@ def test_algorithm_order_is_the_embedded_pair_order(system):
     assert DORMAND_PRINCE_54_TABLEAU.embedded_order == 4
     step = DIRKStep(
         get_solver_helper_fn=system.get_solver_helper,
-        precision=np.float64,
+        precision=np.float32,
         n_states=2,
         tableau=KVAERNO3_TABLEAU,
     )
@@ -74,7 +74,7 @@ def test_smoothed_radau_algorithm_order(system):
 
     step = FIRKStep(
         get_solver_helper_fn=system.get_solver_helper,
-        precision=np.float64,
+        precision=np.float32,
         n_states=2,
         tableau=RADAU_IIA_5_TABLEAU,
         use_smoothed_error=False,
